Enable Lucene Search
This enables the Lucene search, which has been behind a feature flag. The circulars archive now defaults to the `query_string` method, with `multi_match` as a fallback. There is a brief advanced search expandable item on the archive page, and there is an advanced search page in the documentation.